Move settings components to their own pages
