--- src/app/settings/views/Configuration/Security/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Security";
