+++ src/app/settings/views/Security/SecretStorage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SecretStorage";
