+++ src/app/settings/views/Security/SecurityProtocols/SecurityProtocols.tsx
@@ -5,28 +5,21 @@ import { useDispatch, useSelector } from "react-redux";
 
 import TLSDisabled from "./TLSDisabled";
 import TLSEnabled from "./TLSEnabled";
-import VaultSettings from "./VaultSettings";
 
 import { useWindowTitle } from "app/base/hooks";
 import { actions as generalActions } from "app/store/general";
-import {
-  tlsCertificate as tlsCertificateSelectors,
-  vaultEnabled as vaultEnabledSelectors,
-} from "app/store/general/selectors";
+import { tlsCertificate as tlsCertificateSelectors } from "app/store/general/selectors";
 
-const Security = (): JSX.Element => {
+const SecurityProtocols = (): JSX.Element => {
   const dispatch = useDispatch();
   const tlsCertificate = useSelector(tlsCertificateSelectors.get);
   const tlsCertificateLoaded = useSelector(tlsCertificateSelectors.loaded);
-  const vaultEnabledLoaded = useSelector(vaultEnabledSelectors.loaded);
-  useWindowTitle("Security");
-
+  useWindowTitle("Security protocols");
   useEffect(() => {
     dispatch(generalActions.fetchTlsCertificate());
-    dispatch(generalActions.fetchVaultEnabled());
   }, [dispatch]);
 
-  if (!tlsCertificateLoaded && !vaultEnabledLoaded) {
+  if (!tlsCertificateLoaded) {
     return <Spinner text="Loading..." />;
   }
 
@@ -34,7 +27,6 @@ const Security = (): JSX.Element => {
     <>
       <Row>
         <Col size={6}>
-          <h4>Security protocols</h4>
           {!tlsCertificateLoaded ? (
             <Spinner text="Loading..." />
           ) : tlsCertificate ? (
@@ -44,14 +36,8 @@ const Security = (): JSX.Element => {
           )}
         </Col>
       </Row>
-      <Row>
-        <Col size={6}>
-          <h4>Secret storage</h4>
-          <VaultSettings />
-        </Col>
-      </Row>
     </>
   );
 };
 
-export default Security;
+export default SecurityProtocols;
