+++ src/app/settings/views/Security/SecurityProtocols/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SecurityProtocols";
