+++ src/scss/index.scss
@@ -256,7 +256,7 @@
 @import "~app/settings/components/SettingsTable";
 @import "~app/settings/views/Configuration/GeneralForm";
 @import "~app/settings/views/Configuration/GeneralForm/ThemedRadioButton";
-@import "~app/settings/views/Configuration/Security/TLSEnabled/TLSEnabledFields";
+@import "~app/settings/views/Security/SecurityProtocols/TLSEnabled/TLSEnabledFields";
 @import "~app/settings/views/Dhcp/DhcpList";
 @import "~app/settings/views/LicenseKeys/LicenseKeyList";
 @import "~app/settings/views/Repositories/RepositoriesList";
